feat: moved relation checking to property method for better reusability.
